Azken aldaketak
Funtzionatzeari uzten zioten pluginak aldatu.

+++ wp-content/plugins/wpforo/wpf-includes/class-posts.php
@@ -12,8 +12,10 @@ class wpForoPost{
 	}
 	
 	public function add( $args = array() ){
+		
 		if( empty($args) && empty($_REQUEST['post']) ){ $this->wpforo->notice->add('Reply request error', 'error'); return FALSE; }
 		if( empty($args) && !empty($_REQUEST['post']) ){ $args = $_REQUEST['post']; $args['body'] = $_REQUEST['postbody']; }
+		if( !isset($args['body']) || !$args['body'] ){ $this->wpforo->notice->add('Post is empty', 'error'); return FALSE; }
 		
 		extract($args, EXTR_OVERWRITE);
 		
@@ -31,6 +33,8 @@ class wpForoPost{
 			return FALSE;
 		}
 		
+		do_action( 'wpforo_start_add_post', $args );
+		
 		$post = $args;
 		$post['forumid'] = $forumid = (isset($topic['forumid']) ? intval($topic['forumid']) : 0);
 		$post['parentid'] = $parentid = (isset($parentid) ? intval($parentid) : 0);
@@ -40,6 +44,7 @@ class wpForoPost{
 		$post['userid'] = $userid = ( isset($userid) ? intval($userid) : $this->wpforo->current_userid );
 		
 		$post = apply_filters('wpforo_add_post_data_filter', $post);
+		
 		if(empty($post)) return FALSE;
 		
 		extract($post, EXTR_OVERWRITE);
@@ -51,6 +56,7 @@ class wpForoPost{
 		if(isset($created)) $created = sanitize_text_field($created);
 		if(isset($userid)) $userid = intval($userid);
 		if(isset($body)) $body = wpforo_kses(trim($body), 'post');
+		if(isset($status)) $status = intval($status);
 		
 		do_action( 'wpforo_before_add_post', $post );
 		
@@ -65,9 +71,10 @@ class wpForoPost{
 					'title'     => stripslashes($title), 
 					'body'      => stripslashes($body), 
 					'created'	=> $created,
-					'modified'	=> $created
+					'modified'	=> $created,
+					'status'	=> (isset($status) ? $status : 0),
 				), 
-				array('%d','%d','%d','%d','%s','%s','%s','%s')
+				array('%d','%d','%d','%d','%s','%s','%s','%s', '%d')
 			)
 		){
 			$postid = $this->wpforo->db->insert_id;
@@ -101,22 +108,26 @@ class wpForoPost{
 	}
 	
 	public function edit( $args = array() ){
+		
 		if( empty($args) && (!isset($_REQUEST['post']) || empty($_REQUEST['post'])) ) return FALSE;
 		if( empty($args) && !empty($_REQUEST['post']) ){ $args = $_REQUEST['post']; $args['body'] = $_REQUEST['postbody']; }
+		
+		do_action( 'wpforo_start_edit_post', $args );
+		
+		if( !isset($args['postid']) || !$args['postid'] || !is_numeric($args['postid']) ){
+			$this->wpforo->notice->add('Cannot update post data', 'error');
+			return FALSE;
+		}
+		$args['postid'] = intval($args['postid']);
+		if( !$post = $this->get_post($args['postid']) ){ $this->wpforo->notice->add('No Posts found for update', 'error'); return FALSE; }
+		
+		$args['userid'] = $post['userid'];
+		$args['status'] = $post['status'];
 		
 		$args = apply_filters('wpforo_edit_post_data_filter', $args);
 		if(empty($args)) return FALSE;
 		
 		extract($args, EXTR_OVERWRITE);
-		
-		if( !isset($postid) || !$postid || !is_numeric($postid) ){
-			$this->wpforo->notice->add('Cannot update post data', 'error');
-			return FALSE;
-		}
-		
-		$postid = intval($postid);
-		
-		if( !$post = $this->get_post($postid) ){ $this->wpforo->notice->add('No Posts found for update', 'error'); return FALSE; }
 		
 		$diff = current_time( 'timestamp', 1 ) - strtotime($post['created']);
 		if( !($this->wpforo->perm->forum_can('er', $post['forumid']) || ($this->wpforo->current_userid == $post['userid'] && $this->wpforo->perm->forum_can('eor', $post['forumid']) && $diff < $this->wpforo->post_options['eor_durr'])) ){
@@ -135,20 +146,22 @@ class wpForoPost{
 		if(isset($created)) $created = sanitize_text_field($created);
 		if(isset($userid)) $userid = intval($userid);
 		if(isset($body)) $body = wpforo_kses(trim($body), 'post');
+		if(isset($status)) $status = intval($status);
 		
 		if( FALSE !== $this->wpforo->db->update(
 				$this->wpforo->db->prefix."wpforo_posts",
 				array( 
-					'title'      => stripslashes($title), 
-					'body'      => stripslashes($body),
-					'modified'	=> current_time( 'mysql', 1 )
+					'title'     => ( isset($title) ? stripslashes($title) : stripslashes($post['title']) ),
+					'body'      => ( (isset($body) && $body) ? stripslashes($body) : stripslashes($post['body']) ),
+					'modified'	=> current_time( 'mysql', 1 ),
+					'status'  => ( isset($status) ? $status : intval($post['status']) ),
 				), 
 				array('postid' => $postid),
-				array('%s','%s','%s'), 
+				array('%s','%s','%s','%d'), 
 				array('%d') 
 			)
 		){
-			do_action( 'wpforo_after_edit_post', array( 'postid' => $postid, 'topicid' => $topicid, 'title' => $title, 'body' => $body ) );
+			do_action( 'wpforo_after_edit_post', array( 'postid' => $postid, 'topicid' => $topicid, 'title' => $title, 'body' => $body, 'status' => $status ) );
 			
 			$this->wpforo->notice->add('This post successfully edited', 'success');
 			return $postid;
@@ -160,7 +173,7 @@ class wpForoPost{
 	
 	#################################################################################
 	/**
-	 * Delete topic from DB
+	 * Delete post from DB
 	 * 
 	 * Returns true if successfully deleted or false.
 	 *
@@ -173,13 +186,19 @@ class wpForoPost{
 		
 		$postid = intval($postid);
 		
-		$post = $this->get_post($postid);
+		if( !$post = $this->get_post($postid) ) return true;
+		if( $post['is_first_post'] ) return $this->wpforo->topic->delete($post['topicid']);
+
 		$diff = current_time( 'timestamp', 1 ) - strtotime($post['created']);
 		if( !($this->wpforo->perm->forum_can('dr', $post['forumid']) || ($this->wpforo->current_userid == $post['userid'] && $this->wpforo->perm->forum_can('dor', $post['forumid']) && $diff < $this->wpforo->post_options['dor_durr'])) ){
 			$this->wpforo->notice->add('You haven\'t permission to delete post from this forum', 'error');
 			return FALSE;
 		}
 		
+		//Find and delete default atatchments before deleting post
+		$this->delete_attachments( $postid );
+		
+		//Delete post
 		if( $this->wpforo->db->delete($this->wpforo->db->prefix . 'wpforo_posts',  array( 'postid' => intval($postid) ), array( '%d' )) ){
 			$last_post = $this->get_posts( array('topicid' => intval($post['topicid']), 'order' => 'DESC', 'row_count' => 1) );
 			if(is_array($last_post) && !empty($last_post)){
@@ -219,8 +238,13 @@ class wpForoPost{
 			}
 		}
 		
-		$this->wpforo->notice->add('Post delete error', 'error');
-		return FALSE;
+		if( isset($_GET['page']) && $_GET['page'] == 'wpforo-moderations' ){
+			//Don't check errors
+		}
+		else{
+			$this->wpforo->notice->add('Post delete error', 'error');
+			return FALSE;
+		}
 	}
 	
 	#################################################################################
@@ -246,6 +270,12 @@ class wpForoPost{
 		$sql = "SELECT * FROM `".$this->wpforo->db->prefix."wpforo_posts` WHERE `postid` = ".intval($postid);
 		$post = $this->wpforo->db->get_row($sql, ARRAY_A);
 		if(!empty($post)) $post['userid'] = intval($post['userid']);
+		
+		if( isset($post['status']) && $post['status'] && !wpforo_is_owner($post['userid'])){
+			if( isset($post['forumid']) && $post['forumid'] && !$this->wpforo->perm->forum_can('au', $post['forumid']) ){
+				return array();
+			}
+		}
 		
 		if($cache && isset($postid)){
 			self::$cache['post'][$postid] = $post;
@@ -277,7 +307,8 @@ class wpForoPost{
 		  'order'		=> 'ASC', 		// ASC DESC
 		  'offset' 		=> NULL,		// this use when you give row_count
 		  'row_count'	=> NULL, 		// 4 or 1 ...
-		  
+		  'status'		=> NULL, 		    // 0 or 1 ...
+		  		  
 		  'check_private' => FALSE
 		);
 		
@@ -297,9 +328,22 @@ class wpForoPost{
 			if(!empty($include)) $wheres[] = $table_as_prefix . "`postid` IN(" . implode(', ', array_map('intval', $include)) . ")";
 			if(!empty($exclude)) $wheres[] = $table_as_prefix . "`postid` NOT IN(" . implode(', ', array_map('intval', $exclude)) . ")";
 			
-			if($topicid != NULL) $wheres[] = $table_as_prefix . "`topicid` = " . intval($topicid);
+			if(!is_null($topicid)) $wheres[] = $table_as_prefix . "`topicid` = " . intval($topicid);
 			if($parentid != -1) $wheres[]  = $table_as_prefix . "`parentid` = " . intval($parentid);
-			if($userid != NULL) $wheres[]  = $table_as_prefix . "`userid` = " . intval($userid);
+			if(!is_null($userid)) $wheres[]  = $table_as_prefix . "`userid` = " . intval($userid);
+			if(!is_null($status)) $wheres[]  = $table_as_prefix . "`status` = " . intval($status);
+			
+			if( isset($forumid) && $forumid ){
+				if( $this->wpforo->perm->forum_can('au', $forumid) ){
+					if(!is_null($status)) $wheres[] = $table_as_prefix . " `status` = " . intval($status);
+				}
+				elseif( isset($this->wpforo->current_userid) && $this->wpforo->current_userid ){
+					$wheres[] = " ( " . $table_as_prefix .  "`status` = 0 OR (" . $table_as_prefix .  "`status` = 1 AND " . $table_as_prefix .  "`userid` = " .intval($this->wpforo->current_userid). ") )";
+				}
+				else{
+					$wheres[] = " " . $table_as_prefix .  "`status` = 0";
+				}
+			}
 			
 			if( $check_private ){
 				$sql = "SELECT DISTINCT `".$this->wpforo->db->prefix."wpforo_posts`.*, `".$this->wpforo->db->prefix."wpforo_topics`.`private` FROM `".$this->wpforo->db->prefix."wpforo_posts`, `".$this->wpforo->db->prefix."wpforo_topics`";
@@ -342,6 +386,11 @@ class wpForoPost{
 							unset($posts[$key]);
 						}
 					}
+					if( isset($post['forumid']) && isset($post['status']) && $post['status'] && !wpforo_is_owner($post['userid']) ){
+						if( !$this->wpforo->perm->forum_can('au', $post['forumid']) ){
+							unset($posts[$key]);
+						}
+					}
 				}
 			}
 			
@@ -372,7 +421,7 @@ class wpForoPost{
 			
 			$date_period = intval($date_period);
 			
-			$selects = array('p.`postid`', 't.`topicid`', 't.`private`', 't.`forumid`', 'p.`userid`', 't.`title`', 'p.`created`', 'p.`body`' );
+			$selects = array('p.`postid`', 't.`topicid`', 't.`private`', 't.`status`', 't.`forumid`', 'p.`userid`', 't.`title`', 'p.`created`', 'p.`body`' );
 			$innerjoins = array('INNER JOIN `'.$this->wpforo->db->prefix.'wpforo_topics` t ON t.`topicid` = p.`topicid`');
 			$wheres = array();
 			$orders = array();
@@ -432,7 +481,8 @@ class wpForoPost{
 			$posts = $this->wpforo->db->get_results($sql, ARRAY_A);
 			foreach($posts as $key => $post){
 				if( !$this->wpforo->perm->forum_can( 'vf', $post['forumid'] ) ) unset($posts[$key]);
-				if( !$this->wpforo->perm->forum_can( 'vp', $post['forumid'] ) && $post['private'] ) unset($posts[$key]);
+				if( $post['private'] && !$this->wpforo->perm->forum_can( 'vp', $post['forumid'] ) ) unset($posts[$key]);
+				if( $post['status'] && !$this->wpforo->perm->forum_can( 'au', $post['forumid'] ) ) unset($posts[$key]);
 			}
 			return $posts;
 		}else{
@@ -594,14 +644,14 @@ class wpForoPost{
 		}
 		
 		if( is_array($post) && !empty($post) ){
-			if( $cache && isset(self::$cache['post_url'][$post['topicid']][$post['parentid']]) ){
-				$position = self::$cache['post_url'][$post['topicid']][$post['parentid']];
+			if( $cache && isset(self::$cache['post_url'][$post['topicid']][$post['postid']]) ){
+				$position = self::$cache['post_url'][$post['topicid']][$post['postid']];
 			}
 			$url = WPFORO_BASE_URL . $this->get_forumslug_byid($postid) . '/' . $this->get_topicslug_byid($postid);
 			if( $post['topicid'] ){
 				if( !$position ) $position = $this->wpforo->db->get_var("SELECT COUNT(`postid`) FROM `".$this->wpforo->db->prefix."wpforo_posts` WHERE `topicid` = ".intval($post['topicid'])." AND `postid` <= " . ($post['parentid'] ? intval($post['parentid']) : intval($postid) ) . " ORDER BY `postid`");
 				if($cache && isset($position)){
-					self::$cache['post_url'][$post['topicid']][$post['parentid']] = $position;
+					self::$cache['post_url'][$post['topicid']][$post['postid']] = $position;
 				}
 				if( $position <= $this->wpforo->post_options['posts_per_page'] ) return $url . "/#post-" . intval($postid);
 				$paged = ceil( $position/$this->wpforo->post_options['posts_per_page'] );
@@ -638,5 +688,58 @@ class wpForoPost{
 	function get_count(){
 		return $this->wpforo->db->get_var( "SELECT COUNT(`postid`) FROM `".$this->wpforo->db->prefix."wpforo_posts`" );
 	}
+	
+	function unapproved_count(){
+		return $this->wpforo->db->get_var( "SELECT COUNT(*) FROM `".$this->wpforo->db->prefix."wpforo_posts` WHERE `status` = 1" );
+	}
+	
+	function get_attachment_id( $filename ){
+		$attach_id =  $this->wpforo->db->get_var( "SELECT `post_id` FROM `".$this->wpforo->db->prefix."postmeta` WHERE `meta_key` = '_wp_attached_file' AND `meta_value` LIKE '%" . esc_sql($filename) . "' LIMIT 1");
+		return $attach_id;
+	}
+	
+	function delete_attachments( $postid ){
+		$post = $this->get_post($postid);
+		if( isset($post['body']) && $post['body'] ){
+			if( preg_match_all('|\/wpforo\/default_attachments\/([^\s\"\]]+)|is', $post['body'], $attachments, PREG_SET_ORDER) ){
+				$upload_dir = wp_upload_dir();
+                $default_attachments_dir = $upload_dir['basedir'] . '/wpforo/default_attachments/';
+				foreach( $attachments as $attachment ){
+					$filename = trim($attachment[1]);
+					$file = $default_attachments_dir . $filename;
+					if( file_exists($file) ){
+						$posts = $this->wpforo->db->get_var( "SELECT COUNT(*) as posts FROM `".$this->wpforo->db->prefix."wpforo_posts` WHERE `body` LIKE '%" . esc_sql( $attachment[0] ) . "%'" );
+						if( is_numeric($posts) && $posts == 1 ){
+							$attachmentid = $this->get_attachment_id( '/' . $filename  );
+							if ( !wp_delete_attachment( $attachmentid ) ){
+								@unlink($file); 
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+
+	public function status( $postid, $status ){
+        if( !$postid = wpforo_bigintval($postid) ) return false;
+        if( !$post = $this->get_post($postid) ) return false;
+
+        if( $post['is_first_post'] ) return $this->wpforo->topic->status($post['topicid'], $status);
+
+        if( false !== $this->wpforo->db->update(
+            $this->wpforo->db->prefix."wpforo_posts",
+            array( 'status' => intval($status) ),
+            array( 'postid' => $postid ),
+            array( '%d' ),
+            array( '%d' )
+        )){
+            $this->wpforo->notice->add('Done!', 'success');
+            return true;
+        }
+
+        $this->wpforo->notice->add('error: Change Status action', 'error');
+        return false;
+    }
 }
 ?>
